+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 - Pipelines are now tested every time plugins are added/changed/removed
 - Warning badge (!) appears on plugins with missing required fields
 - Copy button in simulation result tooltips to copy JSON data to clipboard
-- Discovery for SNMP
+- Discovery for SNMP - Tested the end-to-end process of discovery and adding a device
+
 
 ### Changed
 - Colorized the "after" json output to be more readable and clearly convey what was changed